requirements.txt
flask
flask-cors
tensorflow-cpu
numpy
opencv-python-headless
pillow
gunicorn
